# backend/showdown-spring/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Showdown Live Spring Backend API
  version: 1.0.0
  description: |
    Showdown Live 시연용 Spring Boot 백엔드 API 명세입니다.

    REST API는 참가, 상점, 관리자 명령, Unreal fallback/debug, H2 DB 확인에 사용합니다.

    실시간 통신은 WebSocket을 사용합니다.

    - Mobile/Host WebSocket: `/ws`
    - Unreal WebSocket: `/ws/unreal`

    OpenAPI는 WebSocket을 정식 endpoint로 표현하지 않으므로, WebSocket payload는 `components.schemas`와 이 설명에 포함합니다.
servers:
  - url: http://localhost:3000
    description: Local Spring server
  - url: https://{cloudflareTunnelHost}
    description: Cloudflare Quick Tunnel URL
    variables:
      cloudflareTunnelHost:
        default: example.trycloudflare.com
...
paths:
...
  /api/join:
    post:
      tags: [Player]
      summary: Join with nickname
      description: Mobile/test client joins the lobby and receives a unique `playerId`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nickname]
              properties:
                nickname:
                  type: string
                  maxLength: 8
                  example: mimi
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResponse"
        "400":
          description: Missing nickname
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/input:
    post:
      tags: [Player]
      summary: HTTP fallback for controller input
      description: Realtime mobile input normally uses `/ws`; this endpoint is for fallback/debug.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InputRequest"
      responses:
        "200":
          description: Input accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  input:
                    $ref: "#/components/schemas/UnrealInput"
        "400":
          description: Game is not Playing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/buy:
    post:
      tags: [Shop]
      summary: Buy one item
      description: A player can buy one item before gameplay starts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, itemId]
              properties:
                playerId:
                  type: string
                  example: 1
                itemId:
                  type: string
                  example: atk_boost
      responses:
        "200":
          description: Purchase succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyResponse"
        "400":
          description: Purchase rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/vote:
    post:
      tags: [Spectator]
      summary: Vote for a spectator event
      description: Only Dead or Spectator players can vote.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, eventType]
              properties:
                playerId:
                  type: string
                  example: 1
                eventType:
                  type: string
                  enum: [HealZone, SpeedUp, ShrinkZone, SupplyBox]
      responses:
        "200":
          description: Vote accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  voteCounts:
                    $ref: "#/components/schemas/VoteCounts"
        "400":
          description: Vote rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cheer:
    post:
      tags: [Spectator]
      summary: Cheer another player
      description: Only Dead or Spectator players can cheer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId, targetId]
              properties:
                playerId:
                  type: string
                  example: 1
                targetId:
                  type: string
                  example: 2
      responses:
        "200":
          description: Cheer accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cheerCounts:
                    $ref: "#/components/schemas/CheerCounts"
        "400":
          description: Cheer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/unreal/player-state:
    post:
      tags: [Unreal]
      summary: Update a player state from Unreal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerStateRequest"
      responses:
        "200":
          description: State updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  player:
                    $ref: "#/components/schemas/Player"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/unreal/positions:
    post:
      tags: [Unreal]
      summary: Upload Unreal player positions for minimap
      description: |
        WebSocket `worldState` is preferred. This REST endpoint remains available.

        Readable payload uses `positions`. Compact payload uses `p`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/PositionsRequest"
                - $ref: "#/components/schemas/CompactPositionsRequest"
      responses:
        "200":
          description: Positions accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  updated:
                    type: integer
                  broadcasted:
                    type: boolean
                  broadcastIntervalMs:
                    type: integer
                    example: 200
  /api/unreal/result:
    post:
      tags: [Unreal]
      summary: Submit final result from Unreal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                winnerId:
                  type: string
                  example: 3
                rankingList:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: Result accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /api/admin/start-shop:
    post:
      tags: [Admin]
      summary: Start shop phase
      responses:
        "200":
          description: Shop started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  gameState:
                    $ref: "#/components/schemas/GameState"
                  shopItems:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShopItem"
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/start-game:
    post:
      tags: [Admin]
      summary: Start gameplay
      responses:
        "200":
          description: Gameplay started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  gameState:
                    $ref: "#/components/schemas/GameState"
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/add-bots:
    post:
      tags: [Admin]
      summary: Add debug bots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                  example: 5
      responses:
        "200":
          description: Bots added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
    GameState:
      type: string
      enum: [Lobby, Shop, Playing, Result]
    PlayerState:
      type: string
      enum: [Joined, Ready, Alive, Dead, Spectator, Winner]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        reason:
          type: string
          example: Player not found
    JoinResponse:
      type: object
      properties:
        success:
          type: boolean
        participantId:
          type: integer
          example: 1
        playerId:
          type: string
          description: String compatibility field. Same numeric value as participantId.
          example: "1"
        nickname:
          type: string
          example: mimi
        state:
          $ref: "#/components/schemas/PlayerState"
    Player:
      type: object
      properties:
        playerId:
          type: string
          description: String compatibility field. Same numeric value as participantId.
          example: "1"
        participantId:
          type: integer
          example: 1
        nickname:
          type: string
          example: mimi
        color:
          type: string
          example: "#a8e6cf"
        state:
          $ref: "#/components/schemas/PlayerState"
        gold:
          type: integer
          example: 60
        items:
          type: array
          items:
            type: string
          example: [atk_boost]
        x:
          type: number
          format: double
          example: 0.5
        y:
          type: number
          format: double
          example: 1.0
        moveX:
          type: number
          format: double
        moveY:
          type: number
          format: double
        jumpSeq:
          type: integer
          format: int64
        emoteSeq:
          type: integer
          format: int64
        posX:
          type: number
          format: double
        posY:
          type: number
          format: double
        worldX:
          type: number
          format: double
        worldY:
          type: number
          format: double
        hp:
          type: number
          format: double
          example: 88
        maxHp:
          type: number
          format: double
          example: 100
        attackPower:
          type: number
          format: double
          example: 1.2
        speedMultiplier:
          type: number
          format: double
          example: 1.0
        rangeMultiplier:
          type: number
          format: double
          example: 1.0
        shield:
          type: number
          format: double
          example: 50
        connected:
          type: boolean
        joinedAt:
          type: integer
          format: int64
        inputTimestamp:
          type: integer
          format: int64
        isBot:
          type: boolean
        vote:
          type: string
          nullable: true
        cheerTargetId:
          type: string
          nullable: true
...
    ShopItem:
      type: object
      properties:
        itemId:
          type: string
          example: atk_boost
        name:
          type: string
          example: 공격 강화
        price:
          type: integer
          example: 40
        stock:
          type: integer
          example: 9
        effect:
          $ref: "#/components/schemas/ItemEffect"
    ItemEffect:
      type: object
      properties:
        type:
          type: string
          enum: [AttackPower, Speed, Range, Shield]
        value:
          oneOf:
            - type: number
            - type: integer
          example: 1.2
    BuyResponse:
      type: object
      properties:
        success:
          type: boolean
        player:
          $ref: "#/components/schemas/Player"
        shopItems:
          type: array
          items:
            $ref: "#/components/schemas/ShopItem"
    VoteCounts:
      type: object
      additionalProperties:
        type: integer
      example:
        HealZone: 1
        SpeedUp: 0
        ShrinkZone: 0
        SupplyBox: 0
    CheerCounts:
      type: object
      additionalProperties:
        type: integer
      example:
        2: 3
    InputRequest:
      type: object
      required: [playerId, moveX, moveY]
      properties:
        playerId:
          type: string
          description: String compatibility field. Same numeric value as participantId.
          example: "1"
        participantId:
          type: integer
          example: 1
        moveX:
          type: number
          format: double
          minimum: -1
          maximum: 1
          example: 0.5
        moveY:
          type: number
          format: double
          minimum: -1
          maximum: 1
          example: 1.0
        jumpSeq:
          type: integer
          format: int64
          example: 3
        emoteSeq:
          type: integer
          format: int64
          example: 1
    UnrealInput:
      type: object
      properties:
        playerId:
          type: string
          description: String compatibility field. Same numeric value as participantId.
          example: "1"
        participantId:
          type: integer
          example: 1
        moveX:
          type: number
          format: double
        moveY:
          type: number
          format: double
        jumpSeq:
          type: integer
          format: int64
        emoteSeq:
          type: integer
          format: int64
        timestamp:
          type: integer
          format: int64
        state:
          $ref: "#/components/schemas/PlayerState"
        connected:
          type: boolean
...
    PlayerStateRequest:
      type: object
      required: [playerId, state]
      properties:
        playerId:
          type: string
          example: 1
        state:
          $ref: "#/components/schemas/PlayerState"
    PositionsRequest:
      type: object
      properties:
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionUpdate"
    PositionUpdate:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: string
          example: 1
        x:
          type: number
          description: Minimap percentage X, 0 to 100
        y:
          type: number
          description: Minimap percentage Y, 0 to 100
        worldX:
          type: number
          description: Unreal world X
        worldY:
          type: number
          description: Unreal world Y
        hp:
          type: number
        maxHp:
          type: number
        alive:
          type: boolean
        state:
          $ref: "#/components/schemas/PlayerState"
    CompactPositionsRequest:
      type: object
      properties:
        p:
          type: array
          description: Compact rows of [playerId, x, y, alive].
          items:
            type: array
            items: {}
          example:
            - [1, 46.6, 48.3, 1]
...